requirements.txt
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
cycler==0.10.0
Cython==0.29.24
et-xmlfile==1.1.0
google-auth==1.35.0
idna==3.2
jdcal==1.4.1
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282420641/work
matplotlib @ file:///tmp/build/80754af9/matplotlib-suite_1613407855456/work
MedPy==0.4.0
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626271506491/work
oauthlib==3.1.1
olefile==0.46
openpyxl @ file:///tmp/build/80754af9/openpyxl_1615411699337/work
pandas==1.2.5
patchify==0.2.3
Pillow @ file:///tmp/build/80754af9/pillow_1625655817137/work
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydensecrf @ git+https://github.com/lucasb-eyer/pydensecrf.git@0d53acbcf5123d4c88040fe68fbb9805fc5b2fb9
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scipy @ file:///tmp/build/80754af9/scipy_1618855647378/work
SimpleITK==2.1.0
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1623709665295/work
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tifffile==2021.8.30
torch==1.9.0
torchaudio==0.9.0a0+33b2469
torchvision==0.10.0
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1624965014186/work
urllib3==1.26.6
Werkzeug==2.0.1
xlrd @ file:///tmp/build/80754af9/xlrd_1608072521494/work
